DiagnosisArena/requirements.txt
python-dotenv>=1.0.1
openai>=1.51.0
anthropic>=0.34.2
google-generativeai>=0.8.3
pandas>=2.2.2
tqdm>=4.66.4
